+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ NEXT (YYYY-MM-DD)
 * OpenTSDB datasource support added (https://github.com/weaveworks/grafanalib/pull/27)
 * Additional fields for ``XAxis``   (https://github.com/weaveworks/grafanalib/pull/28)
 * Add support for Grafana Zabbix plugin (https://github.com/weaveworks/grafanalib/pull/31)
+* Add support for Grafana Zabbix Triggers panel (https://github.com/weaveworks/grafanalib/pull/36)
 
 
 0.1.2 (2017-01-02)